backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-spectacular==0.27.2
flake8==7.1.0
